server/backend/requirements.txt
aniso8601==8.1.0
argon2-cffi==20.1.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JSON==0.3.4
Flask-RESTful==0.3.8
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
neo4j==4.2.1
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
sentry-sdk[flask]==0.19.5
